backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
supabase>=2.8.0
openai>=1.54.0
python-multipart>=0.0.12
pypdf2>=3.0.1
python-docx>=1.1.2
pydantic>=2.9.0
pydantic-settings>=2.5.0
numpy>=1.24.0

